refactor(record): move link and mapper utils to feature-search-lib
