+++ libs/feature/record/src/index.ts
@@ -1,5 +1,2 @@
 export * from './lib/feature-record.module'
 export * from './lib/state'
-export * from './lib/links/link-classifier.service'
-export * from './lib/links/link-utils'
-export * from './lib/mapper'
