+++ libs/feature/record/src/lib/state/mdview.effects.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core'
 import { SearchApiService } from '@geonetwork-ui/data-access/gn4'
-import { ElasticsearchMapper } from '../mapper'
+import { ElasticsearchMapper } from '@geonetwork-ui/feature/search'
 import {
   ElasticsearchService,
   EsSearchResponse,
