+++ libs/feature/search/src/index.ts
@@ -4,3 +4,5 @@ export * from './lib/state/selectors'
 export * from './lib/state/search.facade'
 export * from './lib/state/effects'
 export * from './lib/state/reducer'
+export * from './lib/utils/links'
+export * from './lib/utils/mapper'
