+++ libs/feature/search/src/lib/utils/mapper/atomic-operations.ts
