+++ libs/feature/search/src/lib/utils/mapper/elasticsearch.mapper.spec.ts
