+++ libs/feature/search/src/lib/utils/mapper/elasticsearch.mapper.ts
